// src/app/interfaces/ISlotGroup.ts
//|********************************|| Ascentis Software||******************************|
//* Project        : Puzzle Boards
//*
//* Interface name : IBoardItemGroup
//*
//* Author         : Dean Smith
//* Date created   : 20/05/2021
//*
//* Last Edited by : Dean Smith
//* Date Edited    : 26/05/2021
//*
//* Purpose        : An object to encapsulate loaded slot data from JSON.
//*
//|************************************************************************************|
export interface ISlotGroup{
    amount: number;
    width: number;
    height: number;
}